fix: stop inverting the pool start price twice
The pool add containers already resolve the starting price into the sorted
token0 order, but makeCreatePoolMessageWithApproves inverted it again whenever
the request token pair was out of order. After flipping the token pair the pool
was therefore created at the reciprocal price, which put the selected range
outside the current price and failed the mint slippage check.

Share the starting price resolver between both pool add containers so the value
is always token0-based, and drop the extra inversion from the create pool
message.
