+++ packages/web/src/layouts/pool/pool-add/containers/pool-add-starting-price.utils.ts
@@ -1,0 +1,54 @@
+import BigNumber from "bignumber.js";
+
+import type { TokenModel } from "@models/token/token-model";
+import { checkGnotPath } from "@utils/common";
+import { makeRawPrice } from "@utils/pool-utils";
+import { sortTokenPaths } from "@utils/sort-utils";
+import { priceToNearTick, tickToPrice } from "@utils/swap-utils";
+
+export const snapPoolAddRawStartingPrice = (rawPrice: number, tickSpacing: number): number | null => {
+  if (BigNumber(rawPrice).isNaN() || !Number.isFinite(rawPrice) || rawPrice <= 0) {
+    return null;
+  }
+
+  const tick = priceToNearTick(rawPrice, tickSpacing);
+  return tickToPrice(tick);
+};
+
+const getPoolOrderPath = (token: TokenModel): string => {
+  return token.wrappedPath || checkGnotPath(token.path) || token.path;
+};
+
+/**
+ * The starting price is entered in the order shown on the screen, but a pool always
+ * stores the price based on the sorted token0. Invert it when the base token is token1.
+ */
+export const makePoolAddSortedRawPrice = (
+  displayPrice: number,
+  baseToken: TokenModel,
+  quoteToken: TokenModel,
+): number => {
+  const rawPriceInInputOrder = makeRawPrice(displayPrice, baseToken, quoteToken);
+  const [firstPath] = [getPoolOrderPath(baseToken), getPoolOrderPath(quoteToken)].sort(sortTokenPaths);
+
+  if (firstPath === getPoolOrderPath(baseToken)) {
+    return rawPriceInInputOrder;
+  }
+
+  return BigNumber(1).div(rawPriceInInputOrder).toNumber();
+};
+
+export const resolvePoolAddStartingPrice = (
+  displayPrice: string,
+  tokenA: TokenModel,
+  tokenB: TokenModel,
+  tickSpacing: number,
+): number | null => {
+  const priceNum = BigNumber(displayPrice).toNumber();
+  if (BigNumber(priceNum).isNaN() || !Number.isFinite(priceNum) || priceNum <= 0) {
+    return null;
+  }
+
+  const rawPrice = makePoolAddSortedRawPrice(priceNum, tokenA, tokenB);
+  return snapPoolAddRawStartingPrice(rawPrice, tickSpacing);
+};
