+++ packages/web/src/repositories/pool/pool.message.spec.ts
@@ -14,6 +14,7 @@ jest.mock("@constants/environment.constant", () => ({
 
 import {
   makeCreateExternalIncentiveMessageWithApproves,
+  makeCreatePoolMessageWithApproves,
   makePositionMintMessageWithApproves,
 } from "@repositories/pool/pool.message";
 
@@ -77,6 +78,32 @@ describe("pool.message.ts", () => {
       "1250000",
       "3000000",
     ]);
+  });
+
+  it("uses the given start price as is, whichever order the token pair is passed in", async () => {
+    const caller = "caller";
+    const fetchAllowance = jest.fn(async () => 0);
+    const request = {
+      feeTier: "FEE_3000" as const,
+      startPrice: "50",
+      createPoolFee: 0,
+      caller,
+    };
+
+    const orderedMessages = await makeCreatePoolMessageWithApproves(
+      { ...request, tokenA: createTokenModel("tokenA_path"), tokenB: createTokenModel("tokenB_path") },
+      fetchAllowance,
+    );
+    const reversedMessages = await makeCreatePoolMessageWithApproves(
+      { ...request, tokenA: createTokenModel("tokenB_path"), tokenB: createTokenModel("tokenA_path") },
+      fetchAllowance,
+    );
+
+    const orderedArgs = orderedMessages.find(message => message.func === "CreatePool")?.args;
+    const reversedArgs = reversedMessages.find(message => message.func === "CreatePool")?.args;
+
+    expect(orderedArgs?.slice(0, 3)).toEqual(["tokenA_path", "tokenB_path", "3000"]);
+    expect(reversedArgs).toEqual(orderedArgs);
   });
 
   it("reorders mint token paths and amounts by the sorted token pair", async () => {
